# Improve handling of string IEs