@antoninbas 2023-09-29
New benchmark results:
```
BenchmarkEncodeInfoElementValueToBuffShortString-12    	    2770	    435264 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncodeInfoElementValueToBuffLongString-12     	    1341	    824074 ns/op	       0 B/op	       0 allocs/op
```
Without the use of `copy`:
```
BenchmarkEncodeInfoElementValueToBuffShortString-12    	    2154	    531415 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncodeInfoElementValueToBuffLongString-12     	     152	   7948069 ns/op	       0 B/op	       0 allocs/op
```